feat: api 서비스 수정 및 카카오 로그인 반영
