+++ app/src/main/java/com/example/client/data/api/ApiService.kt
@@ -1,11 +1,16 @@
 package com.example.client.data.api
 
+import com.example.client.data.model.request.KakaoLoginRequest
 import com.example.client.data.model.request.MainOnBoardingRequest
+import retrofit2.Call
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
 
 interface ApiService {
-    @POST("onboarding/submit")
+    @POST("/api/onboarding/save")
     suspend fun submitOnboarding(@Body request: MainOnBoardingRequest): Response<Void>
+
+    @POST("/api/auth/kakao")
+    fun sendKakaoToken(@Body request: KakaoLoginRequest): Call<Void>
 }
