+++ app/src/main/java/com/example/client/data/api/RetrofitClient.kt
@@ -7,11 +7,12 @@ import retrofit2.converter.gson.GsonConverterFactory
 object RetrofitClient {
     private val BASE_URL = BuildConfig.BASE_URL
 
-    val apiService: ApiService by lazy {
-        Retrofit.Builder()
+    val instance: ApiService by lazy {
+        val retrofit = Retrofit.Builder()
             .baseUrl(BASE_URL)
             .addConverterFactory(GsonConverterFactory.create())
             .build()
-            .create(ApiService::class.java)
+
+        retrofit.create(ApiService::class.java)
     }
 }
