+++ app/src/main/java/com/example/client/data/model/request/KakaoLoginRequest.kt
@@ -1,0 +1,5 @@
+package com.example.client.data.model.request
+
+data class KakaoLoginRequest(
+    val accessToken: String
+)
