+++ app/src/main/java/com/example/client/data/model/request/MainOnBoardingRequest.kt
@@ -1,6 +1,7 @@
 package com.example.client.data.model.request
 
 data class MainOnBoardingRequest(
-    val username: String,
-    val answers: List<String>
+    val employmentStatus: String,
+    val region: String,
+    val interestedField: List<String>,
 )
